+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/CategoryController.kt
@@ -2,7 +2,7 @@ package org.shangahi.sellio_backend.api.controller
 
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.shangahi.sellio_backend.api.swagger.doc.CategoryDoc
-import org.shangahi.sellio_backend.api.dto.CategoryRequest
+import org.shangahi.sellio_backend.api.dto.request.CategoryRequest
 import org.shangahi.sellio_backend.api.dto.response.CategoryResponse
 import org.shangahi.sellio_backend.api.mapper.toResponse
 import org.shangahi.sellio_backend.service.CategoryService
