--- src/main/kotlin/org/shangahi/sellio_backend/api/dto/CategoryRequest.kt
@@ -1,3 +1,0 @@
-package org.shangahi.sellio_backend.api.dto
-
-data class CategoryRequest(val title: String)
