+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/request/CategoryRequest.kt
@@ -1,0 +1,3 @@
+package org.shangahi.sellio_backend.api.dto.request
+
+data class CategoryRequest(val title: String)
